Bug of "Cast to ObjectID failed" resolved
I was getting this error:
"Received id: undefined
Cast to ObjectId failed for value "undefined" (type string) at path "_id" for model "Book"
Error in file booksRoute.js @line 32"
in the file booksRoute.js
Upon investigation, found that I mistyped a variable name (Expected_Code -> Error_Code) (book._id -> books_.id)
which was not sending the write URL.
It should have sent:
http://localhost:5173/books/details/64fc2ffc0c0cb2cbe84d7590
But it was sending: http://localhost:5173/books/details/undefined
because of the wrong variable name.
This was a bug that was not easily searchable as the compiler was not able to provide the exact location of the problem. I had to run some small unit tests using "console.log" to figure out the process by breaking the chain of the process and looking at every step.
Now that I have found the bug, I'll check the wrong code in ChatGPT and see if ChatGPT is correctly able to find the error in ONE-PROMPT.

Update: ChatGPT accurately found out the error in one-prompt in which I provided the whole code of files:
booksRoute.js
ShowBooks.jsx
Home.jsx